+++ custom_model_runner/datarobot_drum/drum/common.py
@@ -305,24 +305,6 @@ def verbose_stdout(verbose):
         sys.stdout = old_target
 
 
-@contextmanager
-def capture_R_traceback_if_errors(r_handler, logger):
-    from rpy2.rinterface_lib.embedded import RRuntimeError
-
-    try:
-        yield
-    except RRuntimeError as e:
-        try:
-            out = "\n".join(r_handler("capture.output(traceback(max.lines = 50))"))
-            logger.error("R Traceback:\n{}".format(str(out)))
-        except Exception as traceback_exc:
-            e.context = {
-                "r_traceback": "(an error occurred while getting traceback from R)",
-                "t_traceback_err": traceback_exc,
-            }
-        raise
-
-
 def config_logging():
     logging.basicConfig(format="%(asctime)-15s %(levelname)s %(name)s:  %(message)s")
