+++ custom_model_runner/datarobot_drum/drum/utils.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 
 from scipy.io import mmread
-from contextlib import closing
+from contextlib import closing, contextmanager
 from functools import partial
 from pathlib import Path
 from datarobot_drum.drum.exceptions import DrumCommonException
@@ -286,3 +286,21 @@ class StructuredInputReadUtils:
             raise DrumCommonException(
                 "Pandas failed to read input binary data {}".format(binary_data)
             )
+
+
+@contextmanager
+def capture_R_traceback_if_errors(r_handler, logger):
+    from rpy2.rinterface_lib.embedded import RRuntimeError
+
+    try:
+        yield
+    except RRuntimeError as e:
+        try:
+            out = "\n".join(r_handler("capture.output(traceback(max.lines = 50))"))
+            logger.error("R Traceback:\n{}".format(str(out)))
+        except Exception as traceback_exc:
+            e.context = {
+                "r_traceback": "(an error occurred while getting traceback from R)",
+                "t_traceback_err": traceback_exc,
+            }
+        raise
