+++ custom_model_runner/datarobot_drum/resource/components/Python/r_fit/r_fit.py
@@ -3,8 +3,12 @@ import os
 
 from mlpiper.components.connectable_component import ConnectableComponent
 
-from datarobot_drum.drum.common import LOGGER_NAME_PREFIX, capture_R_traceback_if_errors
-from datarobot_drum.drum.utils import shared_fit_preprocessing, make_sure_artifact_is_small
+from datarobot_drum.drum.common import LOGGER_NAME_PREFIX
+from datarobot_drum.drum.utils import (
+    shared_fit_preprocessing,
+    make_sure_artifact_is_small,
+    capture_R_traceback_if_errors,
+)
 
 logger = logging.getLogger(LOGGER_NAME_PREFIX + "." + __name__)
